requirements.txt
pytest==7.4.2
requests==2.31.0
kafka-python3==3.0.0
discord_notify==1.0.0
spotipy==2.23.0
great-expectations==0.18.13
fastapi
uvicorn[standard]
google-cloud-bigquery
google.cloud-storage